At this point it has become apparent that this project has morphed into so much more than an attempt to clone THQ.
From here on, the stated goal of the project will not be to clone THQ in any way, but rather to create an extensible API and management system.
I do not want it to become a clone of ServiceNow, but I believe it could take a lot of inspiration from it.
As part of this change of heart, I will be focusing the most time towards the API, especially the data validation side of things.
It is my belief that once I establish a solid base for the database and sys_* tables, I will be able to use effectively any front-end that I would like to create. Be it an Angular or React front end.